изменён тип переменных

+++ LR2/Form1.cs
@@ -21,38 +21,38 @@ namespace LR2
         private void button1_Click(object sender, EventArgs e)   // сложение
         {
             
-            int x1 = Convert.ToInt32(textBox1.Text);
-            int x2 = Convert.ToInt32(textBox2.Text);
-            int r = x1 + x2;
+            double x1 = Convert.ToDouble(textBox1.Text);
+            double x2 = Convert.ToDouble(textBox2.Text);
+            double r = x1 + x2;
             textBox3.Text = Convert.ToString(r);
             label4.Text = "Сложение";
         }
 
         private void button2_Click(object sender, EventArgs e)   // вычитание
         {
-            int x1 = Convert.ToInt32(textBox1.Text);
-            int x2 = Convert.ToInt32(textBox2.Text);
-            int r = x1 - x2;
+            double x1 = Convert.ToDouble(textBox1.Text);
+            double x2 = Convert.ToDouble(textBox2.Text);
+            double r = x1 - x2;
             textBox3.Text = Convert.ToString(r);
             label4.Text = "Вычитание";
         }
 
         private void button3_Click(object sender, EventArgs e)   // умножение
         {
-            int x1 = Convert.ToInt32(textBox1.Text);
-            int x2 = Convert.ToInt32(textBox2.Text);
-            int r = x1 * x2;
+            double x1 = Convert.ToDouble(textBox1.Text);
+            double x2 = Convert.ToDouble(textBox2.Text);
+            double r = x1 * x2;
             textBox3.Text = Convert.ToString(r);
             label4.Text = "Умножение";
         }
 
         private void button4_Click(object sender, EventArgs e)   // деление
         {
-            int x1 = Convert.ToInt32(textBox1.Text);
-            int x2 = Convert.ToInt32(textBox2.Text);
+            double x1 = Convert.ToDouble(textBox1.Text);
+            double x2 = Convert.ToDouble(textBox2.Text);
             if (x2 != 0)
             {
-                int r = x1 / x2;
+                double r = x1 / x2;
                 textBox3.Text = Convert.ToString(r);
                 label4.Text = "Деление";
             }
